changed the way data works
data isnt gotten from shared data, its gotten from entity now, whenever entity updates data it updates the shared data instead

+++ Bullet Collab/Assets/Scripts/UIManager.cs
@@ -27,7 +27,7 @@ public class UIManager : MonoBehaviour
     // Update the health bar visual
     public void updateHealth(){
         // Shield Check?
-        int heartMax = dataInfo.maxHealth;
+        float heartMax = dataInfo.maxHealth;
         if (dataInfo.currenthealth > heartMax){
             heartMax = dataInfo.currenthealth;
         }
